refactor!: Organise the hugr-py modules
