+++ hugr-py/src/hugr/__init__.py
@@ -2,8 +2,8 @@
 representation.
 """
 
-from .hugr import Hugr
-from .node_port import Direction, InPort, Node, OutPort, Wire
+from .hugr.base import Hugr
+from .hugr.node_port import Direction, InPort, Node, OutPort, Wire
 from .ops import Op
 from .tys import Kind, Type
