+++ hugr-py/src/hugr/hugr/__init__.py
@@ -1,0 +1,20 @@
+"""The main HUGR structure."""
+
+from .base import Hugr, NodeData
+from .node_port import (
+    Direction,
+    InPort,
+    Node,
+    OutPort,
+    Wire,
+)
+
+__all__ = [
+    "Hugr",
+    "NodeData",
+    "Direction",
+    "InPort",
+    "Wire",
+    "OutPort",
+    "Node",
+]
