+++ hugr-py/tests/test_tracked_dfg.py
@@ -1,10 +1,10 @@
 import pytest
 
 from hugr import tys
+from hugr.build.tracked_dfg import TrackedDfg
 from hugr.ext import Package
 from hugr.std.float import FLOAT_T, FloatVal
 from hugr.std.logic import Not
-from hugr.tracked_dfg import TrackedDfg
 
 from .conftest import CX, QUANTUM_EXT, H, Measure, Rz, validate
